requirements.txt
streamlit
streamlit_extras
google-generativeai
python-dotenv
PyPDF2
Pillow
